% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-subgenre.R
\name{plot_subgenre}
\alias{plot_subgenre}
\title{Generate a Bar Plot of Subgenre Distribution}
\usage{
plot_subgenre(input_genre)
}
\arguments{
\item{input_genre}{The genre of interest for which the subgenre distribution is to be visualized.
Genres must be one of the following: "pop," "rap," "rock," "latin," "r&b," or "edm" (case-sensitive).}
}
\value{
A bar plot displaying the count of each subgenre within the selected genre.
}
\description{
This function takes an input genre and generates a bar plot to visualize the distribution of subgenres
within that genre. It provides insights into the variety of subgenres present in the selected genre.
}
\examples{
# Generate a bar plot for the "rock" genre
plot_subgenre("rock")

}
\seealso{
\code{\link{ggplot}}
}
