% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial-equation.R
\name{polynomial_equation}
\alias{polynomial_equation}
\title{Calculate the Roots of the Polynomial Equation}
\usage{
polynomial_equation(coefficients)
}
\arguments{
\item{coefficients}{A list of numbers depending upon the degree of polynomial equation.
(Combined list of numbers using c(). List should not be more than 6 numbers long meaning
only 5 degrees of freedom allowed. If you want to find roots of a quadratic equation(degrees = 2),
number of coefficients should be degrees+1 i.e. 3)}
}
\value{
Prints a message consisting of roots of the polynomial equation
}
\description{
This function is used to solve polynomial equations, including quadratic equations.

It takes a vector of coefficients as input and determines the nature of the roots
of the polynomial equation. The function provides informative messages indicating
whether the equation has real or complex roots and, if real, how many roots there are.
}
\examples{
# Example 1
quadratic <- c(1,4,4)
polynomial_equation(quadratic)

# Example 2
fourth_order <- c(1, -4, 7, -5, 2)
polynomial_equation(fourth_order)

# Example 3
third_order <- c(1, -4, 7, -5)
polynomial_equation(third_order)

}
\seealso{
\code{\link{polyroot}}
}
\keyword{equation}
\keyword{polynomial}
\keyword{quadratic}
\keyword{roots}
