% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RhythmSync-package.R
\docType{package}
\name{RhythmSync-package}
\alias{RhythmSync}
\alias{RhythmSync-package}
\title{RhythmSync: Rhythms App Launcher and Basic Numeric, Plot Functions}
\description{
The 'RhythmSync' is a package which is designed to make it easy to launch the Rhythms app and also provides a clean spotify songs data from TudyTuesdayR. Alongwith the shinyapp, it also provides a function to solve polynomial equations depending on the number of coefficients provided in the input by the user. An additional function that returns a bar graph of the sub-genres arranged by their count for the selected genre.
}
\author{
\strong{Maintainer}: Harsh Jain \email{hjai0003@student.monash.edu}

}
\keyword{internal}
