% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Runs the Shiny App named Rhythms}
\usage{
run_app()
}
\value{
Shiny App launches in a web browser
}
\description{
Rhythms app takes you through a journey exploring different aspects of music.
It has 5 different chapters each exploring different part of the music data.
Chapter 1: Diving into Diversity: Allows you to explore the distribution of music tracks across different genres and subgenres.
Chapter 2: Dance to the Beat of Audio Features: Helps you uncover the relationships between danceability and energy attributes of music tracks.
Chapter 3: Evolution Across Eras: Allows you to witness the evolution of loudness in music over different time periods.
Chapter 4: Playlists: The Curators of Emotion: Provides insights into the curation of music playlists and how they evoke emotional responses.
Chapter 5: Clustering: Unveiling Hidden Patterns: Helps you visually discern how music tracks are grouped together based on their danceability and energy attributes.
}
\keyword{Shiny}
\keyword{application}
\keyword{data}
\keyword{interactive}
\keyword{music}
\keyword{visualization}
\keyword{web}
